Abstract zpc types to module to include code-gen
This moves the `index.d.ts` into a `types/` directory and renames it to
zapier.custom.d.ts. Then a new `index.d.ts` file has been added, that
imports from both this hand-written new module, and the generated
sibling module, unifying the imports.
With any luck, this commit should include the generated typings because
of the pre-commit hooks.