Introduce schema-to-ts v3, and remove v1 & v2
This takes many of the improvements made in the v2 schema-to-ts
compiler, and refactors them to a radically simpler yet again system.
The plugin system is gone, and a tree of "overrides" is now defined in
constants.ts that lets any type be patched with either a string
replacement, a partial set of options to give to apply to the ts-morph
call, or a function that can take full control of what gets rendered.